Move test code to test/ directory
